r12393/sofa-dev : various warnings fixed
Former-commit-id: 696e7f1004fc79b2c7bb868e3192703b8119914c

+++ modules/sofa/component/forcefield/TetrahedronFEMForceField.h
@@ -195,7 +195,7 @@ public:
                 {
                     for(int i=0; i<3; i++)
                     {
-                        diag->getVector()[e*9 + j*3 + i] = R[j][i];
+                        diag->getVector()[e*9 + j*3 + i] = (float)R[j][i];
                     }
                 }
             }
